Inline NewCourse.vue into NewCourseModal.vue
